requirements.txt
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
dnspython==2.2.1
email-validator==1.1.3
Flask==2.0.3
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
flask-mongoengine==1.0.0
Flask-WTF==1.0.0
idna==3.3
python-dotenv==0.19.2
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.0
mongoengine==0.24.0
pycparser==2.21
pymongo==4.0
requests==2.27.1
six==1.16.0
urllib3==1.26.8
Werkzeug==2.0.3
WTForms==3.0.1
